refactor: Extract Magic Link functions from client.js
