+++ src/libs/client.js
@@ -18,6 +18,10 @@ import {
 
 import packageJSON from '../../package.json'
 
+export {
+  callMagicLinkOnboardingInitClient,
+  connectMagicLinkClient
+} from '/libs/clientHelpers/magicLink'
 export { connectOidcClient } from '/libs/clientHelpers/oidc'
 export { clearClient } from '/libs/clientHelpers/persistClient'
 export { createClient } from '/libs/clientHelpers/createClient'
@@ -396,94 +400,4 @@ export const removeNotificationDeviceToken = async client => {
   })
 
   await saveClient(client)
-}
-
-export const connectMagicLinkClient = async (client, magicCode) => {
-  const stackClient = client.getStackClient()
-
-  let oauthOptions = stackClient.oauthOptions
-  const data = {
-    magic_code: magicCode,
-    client_id: oauthOptions.clientID,
-    client_secret: oauthOptions.clientSecret,
-    scope: '*'
-  }
-
-  const {
-    two_factor_token: twoFactorToken,
-    session_code: sessionCode,
-    ...token
-  } = await stackClient.fetchJSON('POST', '/auth/magic_link/flagship', data)
-
-  const need2FA = twoFactorToken !== undefined
-
-  if (need2FA) {
-    return {
-      client,
-      state: STATE_2FA_NEEDED,
-      twoFactorToken: twoFactorToken
-    }
-  }
-
-  const needFlagshipVerification = sessionCode !== undefined
-
-  if (needFlagshipVerification) {
-    return {
-      client: client,
-      state: STATE_AUTHORIZE_NEEDED,
-      sessionCode: sessionCode
-    }
-  }
-
-  stackClient.setToken(token)
-
-  return {
-    client: client,
-    state: STATE_CONNECTED
-  }
-}
-
-export const callMagicLinkOnboardingInitClient = async ({
-  instance,
-  magicCode
-}) => {
-  const client = await createClient(instance)
-  const stackClient = client.getStackClient()
-
-  await client.certifyFlagship()
-
-  let oauthOptions = stackClient.oauthOptions
-  const data = {
-    magic_code: magicCode,
-    client_id: oauthOptions.clientID,
-    client_secret: oauthOptions.clientSecret,
-    scope: '*'
-  }
-
-  const result = await stackClient.fetchJSON(
-    'POST',
-    '/auth/magic_link/flagship',
-    data
-  )
-
-  if (result.access_token) {
-    stackClient.setToken(result)
-  } else if (result.session_code) {
-    const { session_code } = result
-    const { codeVerifier, codeChallenge } = await createPKCE()
-
-    await client.authorize({
-      sessionCode: session_code,
-      pkceCodes: {
-        codeVerifier,
-        codeChallenge
-      }
-    })
-  }
-
-  await client.login()
-  await saveClient(client)
-  listenTokenRefresh(client)
-
-  return client
 }
